modified:   .vs/pryCalidadDeSoftware/v15/.suo 	modified:   .vs/pryCalidadDeSoftware/v15/Server/sqlite3/storage.ide 	modified:   .vs/pryCalidadDeSoftware/v15/Server/sqlite3/storage.ide-shm 	modified:   .vs/pryCalidadDeSoftware/v15/Server/sqlite3/storage.ide-wal 	modified:   BD/BD.dbmdl 	modified:   BD/BD.jfm 	modified:   CalidadSoftware/Controllers/EmpleadoesController.cs 	modified:   CalidadSoftware/Controllers/HomeController.cs 	modified:   CalidadSoftware/Controllers/PruebaDB.cs 	modified:   CalidadSoftware/Controllers/usersController.cs 	modified:   CalidadSoftware/Models/CalidadSoftware.Context.cs 	modified:   CalidadSoftware/Models/CalidadSoftware.edmx 	modified:   CalidadSoftware/Models/CalidadSoftware.edmx.diagram 	modified:   CalidadSoftware/Models/Empleado.cs 	modified:   CalidadSoftware/Models/users.cs 	modified:   CalidadSoftware/Views/Empleadoes/Index.cshtml 	modified:   CalidadSoftware/Views/Shared/_Layout.cshtml 	modified:   CalidadSoftware/Web.config 	modified:   CalidadSoftware/bin/CalidadSoftware.dll 	modified:   CalidadSoftware/bin/CalidadSoftware.dll.config 	modified:   CalidadSoftware/bin/CalidadSoftware.pdb 	modified:   CalidadSoftware/obj/Debug/CalidadSoftware.dll 	modified:   CalidadSoftware/obj/Debug/CalidadSoftware.pdb 	modified:   CalidadSoftware/obj/Debug/TempPE/Models.CalidadSoftware.Designer.cs.dll 	modified:   CalidadSoftware/obj/Debug/TempPE/Models.CalidadSoftware.cs.dll 	new file:   CalidadSoftware/obj/Debug/TempPE/Models.Model1.Context.cs.dll 	new file:   CalidadSoftware/obj/Debug/TempPE/Models.Model1.Designer.cs.dll 	new file:   CalidadSoftware/obj/Debug/TempPE/Models.Model1.cs.dll 	modified:   CalidadSoftware/obj/Debug/edmxResourcesToEmbed/Models/CalidadSoftware.csdl 	modified:   CalidadSoftware/obj/Debug/edmxResourcesToEmbed/Models/CalidadSoftware.msl 	modified:   CalidadSoftware/obj/Debug/edmxResourcesToEmbed/Models/CalidadSoftware.ssdl

+++ CalidadSoftware/Models/users.cs
@@ -20,10 +20,10 @@ namespace CalidadSoftware.Models
             this.Empleado = new HashSet<Empleado>();
         }
     
-        public decimal id_user { get; set; }
+        public int id_user { get; set; }
         public string user { get; set; }
         public string password { get; set; }
-        public Nullable<decimal> privilege_level { get; set; }
+        public Nullable<int> privilege_level { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Empleado> Empleado { get; set; }
